Switch lib to websocketpp server mode
This change it needed because of the switch to the robot acting as a server.
Switch from client to server allows multiple clients to connect to the robot and track data
It also fixes the unknown ip issue regarding the driver station.
A few additional changes include closing the port after the program gets a SIGTERM and
adding a IterativeRobot class